requirements.txt
playwright==1.41.1
python-dotenv==1.0.0
openai==1.9.0
Pillow==10.1.0
